feat: change file structure cli template react
