+++ code/renderers/react/template/cli/js/Button/Button.stories.js
@@ -1,4 +1,4 @@
-import { Button } from './Button';
+import { Button } from '.';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 export default {
