+++ code/renderers/react/template/cli/js/Header/Header.stories.js
@@ -1,7 +1,6 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { Header } from './Header';
+import { Header } from '.';
 
-const meta = {
+export default {
   title: 'Example/Header',
   component: Header,
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
@@ -10,12 +9,9 @@ const meta = {
     // More on how to position stories at: https://storybook.js.org/docs/react/configure/story-layout
     layout: 'fullscreen',
   },
-} satisfies Meta<typeof Header>;
+};
 
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const LoggedIn: Story = {
+export const LoggedIn = {
   args: {
     user: {
       name: 'Jane Doe',
@@ -23,4 +19,4 @@ export const LoggedIn: Story = {
   },
 };
 
-export const LoggedOut: Story = {};
+export const LoggedOut = {};
