+++ code/renderers/react/template/cli/js/Page/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+import { Header } from '../Header';
 import './page.css';
 
 export const Page = () => {
