+++ code/renderers/react/template/cli/ts-3-8/Button/index.tsx
