+++ code/renderers/react/template/cli/ts-3-8/Header/Header.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { Header } from './Header';
+import { Header } from '.';
 
 const meta: Meta<typeof Header> = {
   title: 'Example/Header',
