+++ code/renderers/react/template/cli/ts-4-9/Header/Header.stories.ts
@@ -1,6 +1,7 @@
-import { Header } from './Header';
+import type { Meta, StoryObj } from '@storybook/react';
+import { Header } from '.';
 
-export default {
+const meta = {
   title: 'Example/Header',
   component: Header,
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
@@ -9,9 +10,12 @@ export default {
     // More on how to position stories at: https://storybook.js.org/docs/react/configure/story-layout
     layout: 'fullscreen',
   },
-};
+} satisfies Meta<typeof Header>;
 
-export const LoggedIn = {
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const LoggedIn: Story = {
   args: {
     user: {
       name: 'Jane Doe',
@@ -19,4 +23,4 @@ export const LoggedIn = {
   },
 };
 
-export const LoggedOut = {};
+export const LoggedOut: Story = {};
